Refactored Master to follow Google Console's package naming conventions. Reset FireBase to follow new naming conventions.
