+++ app/src/androidTest/java/nscad/ad430_5216/family_bank_app/UiTest.java
@@ -1,27 +1,20 @@
-package com.example.family_bank_app;
-
-import android.view.View;
+package nscad.ad430_5216.family_bank_app;
 
 
-import androidx.test.espresso.action.ViewActions;
-
-
+import androidx.test.espresso.matcher.ViewMatchers;
 import androidx.test.ext.junit.rules.ActivityScenarioRule;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
+
+import com.example.family_bank_app.R;
 
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
 import static android.app.PendingIntent.getActivity;
-import static android.service.autofill.Validators.not;
 import static androidx.test.espresso.Espresso.onView;
 import static androidx.test.espresso.action.ViewActions.click;
-import static androidx.test.espresso.action.ViewActions.scrollTo;
-import static androidx.test.espresso.action.ViewActions.typeText;
 import static androidx.test.espresso.assertion.ViewAssertions.matches;
-import static androidx.test.espresso.matcher.RootMatchers.withDecorView;
-import static androidx.test.espresso.matcher.ViewMatchers.hasDescendant;
 import static androidx.test.espresso.matcher.ViewMatchers.hasErrorText;
 import static androidx.test.espresso.matcher.ViewMatchers.isDisplayed;
 import static androidx.test.espresso.matcher.ViewMatchers.withHint;
@@ -46,7 +39,7 @@ public class UiTest {
 
     @Test
     public void hasTextOnScreen() {
-        onView(withId(R.id.MainAccount)).check(matches(withText(R.string.accounts)));
+        onView(ViewMatchers.withId(R.id.MainAccount)).check(matches(withText(R.string.accounts)));
         onView(withId(R.id.toolbar)).check(matches(isDisplayed()));
         onView(withId(R.id.MainLogo)).check(matches(isDisplayed()));
         onView(withId(R.id.CreateAcct)).check(matches(isDisplayed()));
