+++ app/src/main/AndroidManifest.xml
@@ -11,7 +11,7 @@
         android:theme="@style/AppTheme"
         android:requestLegacyExternalStorage="true">
         <activity
-            android:name=".MainActivity"
+            android:name="nscad.ad430_5216.family_bank_app.MainActivity"
             android:label="@string/family_bank"
             android:theme="@style/AppTheme.NoActionBar">
 
@@ -22,12 +22,12 @@
         </activity>
 
         <activity
-            android:name=".AccountActivity"
+            android:name="nscad.ad430_5216.family_bank_app.AccountActivity"
             android:label="@string/account_activity"
             android:theme="@style/AppTheme.NoActionBar"/>
 
         <activity
-            android:name=".TransactionActivity"
+            android:name="nscad.ad430_5216.family_bank_app.TransactionActivity"
             android:label="TransactionActivity"
             android:theme="@style/AppTheme.NoActionBar"/>
     </application>
