+++ app/src/main/java/nscad/ad430_5216/family_bank_app/Dialog_DeleteAccount.java
@@ -1,4 +1,4 @@
-package com.example.family_bank_app;
+package nscad.ad430_5216.family_bank_app;
 
 import android.app.AlertDialog;
 import android.app.Dialog;
@@ -10,9 +10,11 @@ import android.view.View;
 
 import androidx.appcompat.app.AppCompatDialogFragment;
 
-public class Dialog_DeleteTransaction extends AppCompatDialogFragment {
+import com.example.family_bank_app.R;
 
-    private DeleteTransactionDialogListener listener;
+public class Dialog_DeleteAccount extends AppCompatDialogFragment {
+
+    private DeleteAccountDialogListener listener;
 
     @Override
     public Dialog onCreateDialog(Bundle savedInstanceState){
@@ -22,12 +24,12 @@ public class Dialog_DeleteTransaction extends AppCompatDialogFragment {
         View view = inflater.inflate(R.layout.dialog_delete, null);
 
         builder.setTitle("Warning!")
-                .setMessage("Are you sure you wish to delete this transaction?\n\nTHIS CANNOT BE UNDONE!")
+                .setMessage("Are you sure you wish to delete this account and all of its transactions?\n\nTHIS CANNOT BE UNDONE!")
                 .setView(view)
                 .setPositiveButton("Yes", new DialogInterface.OnClickListener() {
                     @Override
                     public void onClick(DialogInterface dialog, int which) {
-                          listener.deleteTransaction(true);
+                        listener.deleteAccount(true);
                     }
                 })
                 .setNegativeButton("No", new DialogInterface.OnClickListener() {
@@ -44,12 +46,12 @@ public class Dialog_DeleteTransaction extends AppCompatDialogFragment {
         super.onAttach(context);
 
         try {
-            listener = (DeleteTransactionDialogListener) context;
+            listener = (DeleteAccountDialogListener) context;
         } catch (ClassCastException e) {
-            throw new ClassCastException(context.toString() + " must implement DeleteTransactionDialogListener");
+            throw new ClassCastException(context.toString() + " must implement DeleteAccountDialogListener");
         }
     }
 
-    public interface DeleteTransactionDialogListener{
-        void deleteTransaction(boolean deleteOK);
+    public interface DeleteAccountDialogListener{
+        void deleteAccount(boolean deleteOK);
     }}
