+++ app/src/main/java/nscad/ad430_5216/family_bank_app/Dialog_DepositWithdraw.java
@@ -1,17 +1,18 @@
-package com.example.family_bank_app;
+package nscad.ad430_5216.family_bank_app;
 
 import android.app.AlertDialog;
 import android.app.Dialog;
 import android.content.Context;
 import android.content.DialogInterface;
 import android.os.Bundle;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.widget.EditText;
 import android.widget.Toast;
 
 import androidx.appcompat.app.AppCompatDialogFragment;
+
+import com.example.family_bank_app.R;
 
 public class Dialog_DepositWithdraw extends AppCompatDialogFragment {
     private static final String TAG = Dialog_DepositWithdraw.class.getSimpleName();
